Debugging black boxes in spectrograms
Added a new debugging file for which the code is tested with just one spectrogram at a time from a working directory set at the top of the debugging file.

Main change in the debugging file was to change the normalization process. Previous process is now commented out and new process is uncommented.

Each spectrogram change/generation is also now saved in working directory so it can be checked easily.